剔除社团下级部门，修改相关逻辑

+++ src/main/java/cn/maxcj/modular/system/controller/DeptController.java
@@ -132,32 +132,10 @@ public class DeptController extends BaseController {
         }
         //完善pids,根据pid拿到pid的pids
         deptSetPids(dept);
-        try {
-            this.deptService.insert(dept);
-            //添加社团下属七个部门
-            for (int i = 0; i < 7; i++){
-                Dept dept1 = new Dept();
-                dept1.setPid(dept.getId());
-                deptSetPids(dept1);
-                switch (i){
-                    case 0 : dept1.setSimplename("秘书处");dept1.setFullname("秘书处"); break;
-                    case 1 : dept1.setSimplename("宣传部");dept1.setFullname("宣传部"); break;
-                    case 2 : dept1.setSimplename("活动部");dept1.setFullname("活动部"); break;
-                    case 3 : dept1.setSimplename("财务部");dept1.setFullname("财务部"); break;
-                    case 4 : dept1.setSimplename("组织部");dept1.setFullname("组织部"); break;
-                    case 5 : dept1.setSimplename("公关部");dept1.setFullname("公关部"); break;
-                    case 6 : dept1.setSimplename("网信部");dept1.setFullname("网信部"); break;
-                    default : break;
-                }
-                this.deptService.insert(dept1);
-            }
-            //初始化该社团对于的社团信息
-            Clubinfo clubinfo = new Clubinfo();
-            clubinfo.setDeptid(dept.getId());
-            clubinfoService.insert(clubinfo);
-        }catch (Exception e){
-            return false;
-        }
+        this.deptService.insert(dept);
+        //初始化社团对应的社团信息
+        Integer id = this.deptService.selectById(dept).getId();
+        clubinfoService.init(id);
         return true;
     }
 
@@ -247,10 +225,10 @@ public class DeptController extends BaseController {
     public Object delete(@RequestParam Integer deptId) {
         //缓存被删除的部门名称
         LogObjectHolder.me().set(ConstantFactory.me().getDeptName(deptId));
-        //查询社团下属部门
-        List<Dept> club_son = this.deptService.club_son(deptId);
-        for (int i = 0; i < club_son.size(); i++){
-            deptService.deleteDept(club_son.get(i).getId());
+        //查询社团下的所有成员，如果有成员不允许删除
+        List<Map<String,Object>> list = this.userService.selectUsers(null,null,null,null, deptId);
+        if (list.size() > 0){
+            return null;
         }
         deptService.deleteDept(deptId);
         return SUCCESS_TIP;
